manual set credit if prefetch is 0

+++ sdk/messaging_servicebus/src/receiver/service_bus_receiver.rs
@@ -20,7 +20,7 @@ where
 
     pub async fn receive_messages(
         &mut self,
-        max_messages: usize,
+        max_messages: u32,
         max_wait_time: Option<std::time::Duration>,
     ) -> Result<Vec<ServiceBusReceivedMessage>, R::ReceiveError> {
         self.inner
